refactor: ♻️ move expense EIP-712 helpers into expenseUtil
budgetLimitTypes and buildContractBudgetLimit are pure translations of a
budget limit into what the contract expects, with no ties to the
component that happened to hold them — and TransferAction had grown past
the 300-line lint cap.
